Revert "service discovery: Reduce overhead of readlink calls (#36420)"
This reverts commit ded58c04bc47824ec3647bc96b4fe6f9d6f8c51d.
